BeibeiamigosGPT/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
langdetect==1.0.9
MarkupSafe==3.0.2
openai==1.74.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
